requirements.txt
customtkinter==5.2.2
python-dotenv==1.0.1
pillow==10.4.0
requests==2.32.3
pyvista==0.44.0  # STL Viewer
reportlab==4.2.2 # PDF Reports
